+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dto/PendingTestParticipantDTO.java
@@ -6,9 +6,10 @@ public class PendingTestParticipantDTO {
 	private Long id;
 	private String uniqueId;
 	private String gender;
-	private Integer age;
 	private Long educationTypeId;
 	private EducationTypeDTO educationType;
+	private Long ageRangeId;
+	private AgeRangeDTO ageRange;
 	private String occupation;
 	private Integer professionalExperienceMonths;
 	private Integer computerExperienceMonths;
@@ -20,11 +21,14 @@ public class PendingTestParticipantDTO {
 	public PendingTestParticipantDTO(PendingTestParticipantEntity entity) {
 		this.setId(entity.getId());
 		this.uniqueId = entity.getUniqueId();
-		this.age = entity.getAge();
 		this.gender = entity.getGender();
 		this.educationTypeId = entity.getEducationTypeId();
 		if(entity.getEducation() != null) {
 			this.educationType = new EducationTypeDTO(entity.getEducation());
+		}
+		this.ageRangeId = entity.getAgeRangeId();
+		if(entity.getAgeRange() != null) {
+			this.ageRange = new AgeRangeDTO(entity.getAgeRange());
 		}
 		this.occupation = entity.getOccupation();
 		this.professionalExperienceMonths = entity.getProfessionalExperienceMonths();
@@ -55,15 +59,7 @@ public class PendingTestParticipantDTO {
 	public void setGender(String gender) {
 		this.gender = gender;
 	}
-
-	public Integer getAge() {
-		return age;
-	}
-
-	public void setAge(Integer age) {
-		this.age = age;
-	}
-
+	
 	public Long getEducationTypeId() {
 		return educationTypeId;
 	}
@@ -118,5 +114,21 @@ public class PendingTestParticipantDTO {
 
 	public void setEducationType(EducationTypeDTO educationType) {
 		this.educationType = educationType;
+	}
+
+	public Long getAgeRangeId() {
+		return ageRangeId;
+	}
+
+	public void setAgeRangeId(Long ageRangeId) {
+		this.ageRangeId = ageRangeId;
+	}
+
+	public AgeRangeDTO getAgeRange() {
+		return ageRange;
+	}
+
+	public void setAgeRange(AgeRangeDTO ageRange) {
+		this.ageRange = ageRange;
 	}
 }