+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/entity/CertificationCriterionEntity.java
@@ -24,7 +24,7 @@ import javax.persistence.Transient;
  */
 
 @Entity
-@Table(name = "certification_criterion", schema = "openchpl")
+@Table(name = "certification_criterion")
 public class CertificationCriterionEntity implements Serializable {
 
 	/** Serial Version UID. */