+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/entity/TestParticipantEntity.java
@@ -17,9 +17,7 @@ import javax.validation.constraints.NotNull;
 
 @Entity
 @Table(name = "test_participant")
-public class TestParticipantEntity implements Cloneable, Serializable {
-	private static final long serialVersionUID = -606995659042551741L;
-
+public class TestParticipantEntity {
 	@Id 
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Basic( optional = false )
@@ -28,16 +26,20 @@ public class TestParticipantEntity implements Cloneable, Serializable {
 
 	@Column(name = "gender")
 	private String gender;
-	
-	@Column(name = "age" )
-	private Integer age;
 
-	@Column( name = "education_type_id", nullable = false  )
+	@Column( name = "education_type_id", nullable = true  )
 	private Long educationTypeId;
 	
 	@OneToOne(optional = true, fetch = FetchType.LAZY)
-	@JoinColumn(name = "education_type_id", unique=true, nullable = true, insertable=false, updatable=false)
+	@JoinColumn(name = "education_type_id", nullable = true, insertable = false, updatable = false)
 	private EducationTypeEntity education;
+	
+	@Column( name = "test_participant_age_id", nullable = true )
+	private Long ageRangeId;
+	
+	@OneToOne(optional = true, fetch = FetchType.LAZY)
+	@JoinColumn(name = "test_participant_age_id", nullable = true, insertable = false, updatable = false)
+	private AgeRangeEntity ageRange;
 	
 	@Column(name = "occupation")
 	private String occupation;
@@ -87,14 +89,6 @@ public class TestParticipantEntity implements Cloneable, Serializable {
 
 	public void setGender(String gender) {
 		this.gender = gender;
-	}
-
-	public Integer getAge() {
-		return age;
-	}
-
-	public void setAge(Integer age) {
-		this.age = age;
 	}
 
 	public Long getEducationTypeId() {
@@ -183,5 +177,21 @@ public class TestParticipantEntity implements Cloneable, Serializable {
 
 	public void setLastModifiedUser(Long lastModifiedUser) {
 		this.lastModifiedUser = lastModifiedUser;
+	}
+
+	public Long getAgeRangeId() {
+		return ageRangeId;
+	}
+
+	public void setAgeRangeId(Long ageRangeId) {
+		this.ageRangeId = ageRangeId;
+	}
+
+	public AgeRangeEntity getAgeRange() {
+		return ageRange;
+	}
+
+	public void setAgeRange(AgeRangeEntity ageRange) {
+		this.ageRange = ageRange;
 	}
 }